Move v1 model fields from Journalpost to AbstraktRegistrering base type
We move received date, sent date, and public assessment date because
we want to create a new Arkivnotat object that will need to have the
same fields. The fields remain optional for basic record and meeting
record.
